Remove old mongodb logger
