+++ backend/TwitchLogger/main.js
@@ -1,18 +1,28 @@
-const MongoClient = require('mongodb').MongoClient;
+const {Pool, Client} = require('pg');
+const tmi = require('tmi.js');
 const fs = require('fs');
 
 const twitchLogging = require('./twitchLogging.js');
 
-const dbOptions = JSON.parse(fs.readFileSync('mongodbSettings.json'));
-const url = `mongodb://${dbOptions.username}:${dbOptions.password}@${dbOptions.address}:${dbOptions.port}/?compressors=zlib`;
+// read config files
+const options = JSON.parse(fs.readFileSync('twitchSettings.json'));
+const dbOptions = JSON.parse(fs.readFileSync('sqlSettings.json'));
 
-MongoClient.connect(url, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true
-}, (err, client) => {
-    if(err) {
-        console.error(err);
-        return;
-    }
-    twitchLogging.start(client);
+// connect to database
+const sqlPool = new Pool({
+    user: dbOptions.user,
+    host: dbOptions.host,
+    database: dbOptions.database,
+    password: dbOptions.password,
+    port: dbOptions.port
 });
+sqlPool.connect();
+
+// connect to twitch
+client = new tmi.client(options);
+client.connect();
+
+// start logging
+twitchLogging.start(sqlPool, client);
+
+
