--- backend/TwitchLoggerSQL/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "twitchloggersql",
-  "version": "1.0.0",
-  "description": "",
-  "main": "main.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "pg": "^8.2.1",
-    "tmi.js": "^1.5.0"
-  }
-}
